+++ src/main/java/org/stellar/sdk/SorobanServer.java
@@ -489,7 +489,7 @@ public class SorobanServer implements Closeable {
    *     org.stellar.sdk.operations.RestoreFootprintOperation}. Any provided footprint will be
    *     ignored.
    * @param resourceConfig Additional resource include in the simulation.
-   * @param authMode
+   * @param authMode Explicitly allows users to opt-in to non-root authorization in recording mode.
    * @return A {@link SimulateTransactionResponse} object containing the cost, footprint,
    *     result/auth requirements (if applicable), and error of the transaction.
    * @throws org.stellar.sdk.exception.NetworkException All the exceptions below are subclasses of
@@ -503,10 +503,12 @@ public class SorobanServer implements Closeable {
    *     target="_blank">simulateTransaction documentation</a>
    */
   public SimulateTransactionResponse simulateTransaction(
-      Transaction transaction, @Nullable SimulateTransactionRequest.ResourceConfig resourceConfig, @Nullable SimulateTransactionAuthMode authMode) {
+      Transaction transaction,
+      @Nullable SimulateTransactionRequest.ResourceConfig resourceConfig,
+      @Nullable SimulateTransactionRequest.AuthMode authMode) {
     // TODO: In the future, it may be necessary to consider FeeBumpTransaction.
     SimulateTransactionRequest params =
-        new SimulateTransactionRequest(transaction.toEnvelopeXdrBase64(), resourceConfig);
+        new SimulateTransactionRequest(transaction.toEnvelopeXdrBase64(), resourceConfig, authMode);
     return this.sendRequest(
         "simulateTransaction",
         params,
@@ -515,10 +517,11 @@ public class SorobanServer implements Closeable {
 
   /**
    * An alias for {@link #simulateTransaction(Transaction,
-   * SimulateTransactionRequest.ResourceConfig)} with no resource leeway.
+   * SimulateTransactionRequest.ResourceConfig, SimulateTransactionRequest.AuthMode)} with no
+   * resource leeway, and no auth mode set.
    */
   public SimulateTransactionResponse simulateTransaction(Transaction transaction) {
-    return simulateTransaction(transaction, null);
+    return simulateTransaction(transaction, null, null);
   }
 
   /**
@@ -819,24 +822,5 @@ public class SorobanServer implements Closeable {
   public interface SleepStrategy extends IntFunction<Long> {
     // apply as in apply(iterationNumber) -> millisecondsToSleep
     Long apply(int iteration);
-  }
-
-
-  /**
-   * Defines the authorization mode for simulating transactions.
-   */
-  public enum SimulateTransactionAuthMode {
-    /**
-     * Always enforce mode, even with an empty list.
-     */
-    ENFORCE,
-    /**
-     * Always recording mode, failing if any auth exists.
-     */
-    RECORD,
-    /**
-     * Always recording mode, failing if any auth exists, but allowing non-root authorization.
-     */
-    RECORD_ALLOW_NONROOT;
   }
 }