+++ src/main/java/org/stellar/sdk/requests/sorobanrpc/SimulateTransactionRequest.java
@@ -1,5 +1,6 @@
 package org.stellar.sdk.requests.sorobanrpc;
 
+import com.google.gson.annotations.SerializedName;
 import java.math.BigInteger;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -18,7 +19,25 @@ public class SimulateTransactionRequest {
 
   ResourceConfig resourceConfig;
 
+  /**
+   * Explicitly allows users to opt-in to non-root authorization in recording mode.
+   *
+   * <p>Leaving this field unset will default to {@link AuthMode#ENFORCE} if auth entries are
+   * present, {@link AuthMode#RECORD} otherwise.
+   */
+  AuthMode authMode;
 
+  public enum AuthMode {
+    /** Always enforce mode, even with an empty list. */
+    @SerializedName("enforce")
+    ENFORCE,
+    /** Always recording mode, failing if any auth exists. */
+    @SerializedName("record")
+    RECORD,
+    /** Like {@code RECORD}, but allowing non-root authorization. */
+    @SerializedName("record_allow_nonroot")
+    RECORD_ALLOW_NONROOT;
+  }
 
   @Value
   @AllArgsConstructor