+++ src/test/java/org/stellar/sdk/SorobanServerTest.java
@@ -998,7 +998,79 @@ public class SorobanServerTest {
 
     SimulateTransactionRequest.ResourceConfig resourceConfig =
         new SimulateTransactionRequest.ResourceConfig(cpuInstructions);
-    SimulateTransactionResponse resp = server.simulateTransaction(transaction, resourceConfig);
+    SimulateTransactionResponse resp =
+        server.simulateTransaction(transaction, resourceConfig, null);
+    assertEquals(resp.getLatestLedger().longValue(), 14245L);
+    assertEquals(
+        resp.getTransactionData(),
+        "AAAAAAAAAAIAAAAGAAAAAem354u9STQWq5b3Ed1j9tOemvL7xV0NPwhn4gXg0AP8AAAAFAAAAAEAAAAH8dTe2OoI0BnhlDbH0fWvXmvprkBvBAgKIcL9busuuMEAAAABAAAABgAAAAHpt+eLvUk0FquW9xHdY/bTnpry+8VdDT8IZ+IF4NAD/AAAABAAAAABAAAAAgAAAA8AAAAHQ291bnRlcgAAAAASAAAAAAAAAABYt8SiyPKXqo89JHEoH9/M7K/kjlZjMT7BjhKnPsqYoQAAAAEAHifGAAAFlAAAAIgAAAAAAAAAAg==");
+    assertEquals(resp.getEvents().size(), 2);
+    assertEquals(
+        resp.getEvents().get(0),
+        "AAAAAQAAAAAAAAAAAAAAAgAAAAAAAAADAAAADwAAAAdmbl9jYWxsAAAAAA0AAAAg6bfni71JNBarlvcR3WP2056a8vvFXQ0/CGfiBeDQA/wAAAAPAAAACWluY3JlbWVudAAAAAAAABAAAAABAAAAAgAAABIAAAAAAAAAAFi3xKLI8peqjz0kcSgf38zsr+SOVmMxPsGOEqc+ypihAAAAAwAAAAo=");
+    assertEquals(
+        resp.getEvents().get(1),
+        "AAAAAQAAAAAAAAAB6bfni71JNBarlvcR3WP2056a8vvFXQ0/CGfiBeDQA/wAAAACAAAAAAAAAAIAAAAPAAAACWZuX3JldHVybgAAAAAAAA8AAAAJaW5jcmVtZW50AAAAAAAAAwAAABQ=");
+    assertEquals(resp.getMinResourceFee().longValue(), 58181L);
+    assertEquals(resp.getResults().size(), 1);
+    assertEquals(resp.getResults().get(0).getAuth().size(), 1);
+    assertEquals(
+        resp.getResults().get(0).getAuth().get(0),
+        "AAAAAAAAAAAAAAAB6bfni71JNBarlvcR3WP2056a8vvFXQ0/CGfiBeDQA/wAAAAJaW5jcmVtZW50AAAAAAAAAgAAABIAAAAAAAAAAFi3xKLI8peqjz0kcSgf38zsr+SOVmMxPsGOEqc+ypihAAAAAwAAAAoAAAAA");
+    assertEquals(resp.getResults().get(0).getXdr(), "AAAAAwAAABQ=");
+    server.close();
+    mockWebServer.close();
+  }
+
+  @Test
+  public void testSimulateTransactionWithAuthMode() throws IOException, SorobanRpcException {
+    String filePath =
+        "src/test/resources/soroban_server/simulate_transaction_with_resource_leeway.json";
+    String json = new String(Files.readAllBytes(Paths.get(filePath)));
+    Transaction transaction = buildSorobanTransaction(null, null);
+    BigInteger cpuInstructions = BigInteger.valueOf(20000L);
+    MockWebServer mockWebServer = new MockWebServer();
+    Dispatcher dispatcher =
+        new Dispatcher() {
+          @NotNull
+          @Override
+          public MockResponse dispatch(@NotNull RecordedRequest recordedRequest)
+              throws InterruptedException {
+            SorobanRpcRequest<SimulateTransactionRequest> sorobanRpcRequest =
+                gson.fromJson(
+                    recordedRequest.getBody().readUtf8(),
+                    new TypeToken<SorobanRpcRequest<SimulateTransactionRequest>>() {}.getType());
+            if ("POST".equals(recordedRequest.getMethod())
+                && sorobanRpcRequest.getMethod().equals("simulateTransaction")
+                && sorobanRpcRequest
+                    .getParams()
+                    .getTransaction()
+                    .equals(transaction.toEnvelopeXdrBase64())
+                && sorobanRpcRequest
+                    .getParams()
+                    .getResourceConfig()
+                    .getInstructionLeeway()
+                    .equals(cpuInstructions)
+                && sorobanRpcRequest
+                    .getParams()
+                    .getAuthMode()
+                    .equals(SimulateTransactionRequest.AuthMode.RECORD_ALLOW_NONROOT)) {
+              return new MockResponse().setResponseCode(200).setBody(json);
+            }
+            return new MockResponse().setResponseCode(404);
+          }
+        };
+    mockWebServer.setDispatcher(dispatcher);
+    mockWebServer.start();
+
+    HttpUrl baseUrl = mockWebServer.url("");
+    SorobanServer server = new SorobanServer(baseUrl.toString());
+
+    SimulateTransactionRequest.ResourceConfig resourceConfig =
+        new SimulateTransactionRequest.ResourceConfig(cpuInstructions);
+    SimulateTransactionResponse resp =
+        server.simulateTransaction(
+            transaction, resourceConfig, SimulateTransactionRequest.AuthMode.RECORD_ALLOW_NONROOT);
     assertEquals(resp.getLatestLedger().longValue(), 14245L);
     assertEquals(
         resp.getTransactionData(),